+++ src/main/java/uk/gov/pay/connector/service/search/SearchService.java
@@ -2,7 +2,7 @@ package uk.gov.pay.connector.service.search;
 
 import uk.gov.pay.connector.dao.CardTypeDao;
 import uk.gov.pay.connector.dao.ChargeDao;
-import uk.gov.pay.connector.dao.TransactionDao;
+import uk.gov.pay.connector.dao.OldTransactionDao;
 import uk.gov.pay.connector.service.ChargeService;
 
 import javax.inject.Inject;
@@ -12,14 +12,14 @@ public class SearchService {
     public enum TYPE {CHARGE, TRANSACTION}
 
     private ChargeDao chargeDao;
-    private TransactionDao transactionDao;
+    private OldTransactionDao oldTransactionDao;
     private ChargeService chargeService;
     private final CardTypeDao cardTypeDao;
 
     @Inject
-    public SearchService(ChargeDao chargeDao, TransactionDao transactionDao, ChargeService chargeService, CardTypeDao cardTypeDao) {
+    public SearchService(ChargeDao chargeDao, OldTransactionDao oldTransactionDao, ChargeService chargeService, CardTypeDao cardTypeDao) {
         this.chargeDao = chargeDao;
-        this.transactionDao = transactionDao;
+        this.oldTransactionDao = oldTransactionDao;
         this.chargeService = chargeService;
         this.cardTypeDao = cardTypeDao;
     }
@@ -27,7 +27,7 @@ public class SearchService {
     public SearchStrategy ofType(TYPE type) {
         switch (type) {
             case TRANSACTION:
-                return new TransactionSearchStrategy(transactionDao, cardTypeDao);
+                return new OldTransactionSearchStrategy(oldTransactionDao, cardTypeDao);
             default:
                 return new ChargeSearchStrategy(chargeService, chargeDao, cardTypeDao);
         }
