+++ src/test/java/uk/gov/pay/connector/it/dao/GuicedTestEnvironment.java
@@ -6,7 +6,13 @@ import com.google.inject.Injector;
 import com.google.inject.Singleton;
 import com.google.inject.persist.PersistService;
 import com.google.inject.persist.jpa.JpaPersistModule;
-import uk.gov.pay.connector.dao.*;
+import uk.gov.pay.connector.dao.CardDao;
+import uk.gov.pay.connector.dao.ChargeDao;
+import uk.gov.pay.connector.dao.ChargeEventDao;
+import uk.gov.pay.connector.dao.GatewayAccountDao;
+import uk.gov.pay.connector.dao.PaymentRequestDao;
+import uk.gov.pay.connector.dao.TokenDao;
+import uk.gov.pay.connector.dao.TransactionDao;
 
 public class GuicedTestEnvironment {
 
@@ -44,7 +50,7 @@ public class GuicedTestEnvironment {
             bind(GatewayAccountDao.class).in(Singleton.class);
             bind(PaymentRequestDao.class).in(Singleton.class);
             bind(CardDao.class).in(Singleton.class);
-            bind(NewTransactionDao.class).in(Singleton.class);
+            bind(TransactionDao.class).in(Singleton.class);
         }
     }
 }
